Add section image type variable to project pages frontmatter
Currently there are two types of image arrays that can be added to
project page sections, image-full and image-grid. The first tells the
depleting engine to just show each image at full available width on the
page. This type is to be used for a smallest number of images. The
second variable, image-grid is to be used when wanting to show a larger
number of images. The images in these section will be displayed in a
grid format. This variable is best used when each individual image is
not as important as the grouping of images in aiding the writer
describe or show a concept like process work consisting of rough
sketches on paper.